patch up remaining refactored modules
