+++ src/JlWrap/callback.jl
@@ -111,7 +111,6 @@ function pyfunc(
     end
     return f3
 end
-export pyfunc
 
 """
     pyclassmethod(f; ...)
@@ -123,7 +122,6 @@ If `f` is not a Python object (e.g. if `f` is a `Function`) then it is converted
 `Py`. Keyword arguments are passed to `pyfunc`.
 """
 pyclassmethod(f; kw...) = pybuiltins.classmethod(ispy(f) ? f : pyfunc(f; kw...))
-export pyclassmethod
 
 """
     pystaticmethod(f; ...)
@@ -135,7 +133,6 @@ If `f` is not a Python object (e.g. if `f` is a `Function`) then it is converted
 `Py`. Any keyword arguments are passed to `pyfunc`.
 """
 pystaticmethod(f; kw...) = pybuiltins.staticmethod(ispy(f) ? f : pyfunc(f; kw...))
-export pystaticmethod
 
 """
     pyproperty(; get=nothing, set=nothing, del=nothing, doc=nothing)
@@ -155,4 +152,3 @@ pyproperty(; get = nothing, set = nothing, del = nothing, doc = nothing) =
         doc = doc,
     )
 pyproperty(get) = pyproperty(get = get)
-export pyproperty
