+++ src/JlWrap/io.jl
@@ -337,7 +337,6 @@ Wrap `io` as a Python binary IO object.
 This is the default behaviour of `Py(io)`.
 """
 pybinaryio(v::IO) = pyjl(pyjlbinaryiotype, v)
-export pybinaryio
 
 """
     pytextio(io::IO)
@@ -345,6 +344,5 @@ export pybinaryio
 Wrap `io` as a Python text IO object.
 """
 pytextio(v::IO) = pyjl(pyjltextiotype, v)
-export pytextio
 
 pyjltype(::IO) = pyjlbinaryiotype
